+++ cdlclient/tests/remoteclient_unit.py
@@ -65,7 +65,8 @@ def multiple_commands(remote: SimpleRemoteProxy):
         remote.reset_all()
         remote.open_h5_files([fname], True, False)
         remote.import_h5_file(fname, True)
-        remote.switch_to_panel("signal")
+        remote.set_current_panel("signal")
+        assert remote.get_current_panel() == "signal"
         remote.calc("log10")
 
         remote.calc("fft")