+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changes:
 
 * Remote API:
   * Added dict-like interface (e.g. `proxy['obj123']`)
+  * Renamed `switch_to_panel` method to `set_current_panel` (compatibility with DataLab 0.9)
+  * Added `get_current_panel` method
 
 ## Version 0.3.0 ##