+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes:
   * Renamed `switch_to_panel` method to `set_current_panel` (compatibility with DataLab 0.9)
   * Added `get_current_panel` method
   * Changed `select_groups` first argument `selection` (compatibility with DataLab 0.9)
+  * Changed `select_objects` arguments (compatibility with DataLab 0.9)
 
 ## Version 0.3.0 ##